gb/test: update for go1.8 build tag and testing.testmain API change
