+++ test/goversion14_test.go
@@ -1,5 +1,5 @@
 // +build go1.4
-// +build !go1.5,!go1.6
+// +build !go1.5,!go1.6,!go1.7
 
 package test
