+++ test/goversion17_test.go
@@ -1,8 +1,7 @@
 // +build go1.6
-// +build !go1.7
 
 package test
 
 const (
-	goversion = 1.6
+	goversion = 1.7
 )
